pyproject.toml
[project]
name = "academic-research-mentor"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "MajorTimberWolf", email = "ak0929@srmist.edu.in" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.2",
    "langchain-openai>=0.1.22",
    "langchain-google-genai>=2.0.6",
    "langchain-anthropic>=0.2.4",
    "langchain-mistralai>=0.2.1",
    "langchain-community>=0.2",
    "google-genai>=1.29.0",
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "mistralai>=0.4.2",
    "httpx>=0.27.0",
    "langgraph>=0.6.5",
    "rich>=14.1.0",
    "python-dotenv>=1.0.0",
    "ddgs>=9.5.5",
    "beautifulsoup4>=4.12.0",
    "langchain-text-splitters>=0.2",
    "faiss-cpu>=1.12.0",
    "pypdf>=6.0.0",
    "fastembed>=0.7.3",
    "pymupdf>=1.26.4",
    "tavily-python>=0.3.3",
]

[project.scripts]
academic-research-mentor = "academic_research_mentor.cli:main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
]
